Fix install checking when installed from UnityHub
Seems that UnityHub may have issues with sanitising its own data so
the `installedSize` for a package would sometimes be a float:

   "downloadSize": 291803386,
   "installedSize": 595278907.44,

which would throw an error parsing it to an int64.

As its not used, we can just not read it and continue on.

Additional documentation has been added to help improve usage